+++ changelog.md
@@ -13,5 +13,7 @@
 
 ## Improvements
 
+* Added a new paramter `undirected_relationship_types` to `gds.alpha.graph.construct` which allows constructing undirected graphs, when using GDS >= 2.3.0. 
+
 
 ## Other changes